Feature/measurement

+++ main_server/src/main/kotlin/ac/kr/smu/prlab_server/service/impl/MeasurementDataServiceImpl.kt
@@ -8,11 +8,13 @@ import ac.kr.smu.prlab_server.enums.Expression
 import ac.kr.smu.prlab_server.enums.MeasurementTarget
 import ac.kr.smu.prlab_server.enums.Metric
 import ac.kr.smu.prlab_server.enums.Period
+import ac.kr.smu.prlab_server.exception.MeasurementDataPermissionException
 import ac.kr.smu.prlab_server.repository.MeasurementDataRepository
 import ac.kr.smu.prlab_server.service.MeasurementDataService
 import ac.kr.smu.prlab_server.util.*
 import org.springframework.data.domain.PageRequest
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDateTime
 import java.time.LocalTime
 import java.time.ZoneId
@@ -25,18 +27,31 @@ import kotlin.jvm.optionals.getOrNull
 class MeasurementDataServiceImpl(
     private val repo: MeasurementDataRepository
 ): MeasurementDataService {
+
+    @Transactional(readOnly = true)
     override fun findRecentData(userId:Long): RecentData {
         return repo.findRecentData(userId)
     }
 
-    override fun findById(id: Long): MeasurementData? {
-        return repo.findById(id).getOrNull()
+    @Transactional(readOnly = true)
+    override fun findById(userId: Long, id: Long): MeasurementData? {
+        val data = repo.findById(id).getOrNull() ?: return null
+        if(data.user.uid != userId)
+            throw MeasurementDataPermissionException()
+        return data
     }
+//    @Transactional(readOnly = true)
+    override fun deleteById(userID: Long, id: Long) {
+        val data = repo.findById(id).getOrNull()
 
-    override fun deleteById(id: Long) {
+        checkNotNull(data){"데이터가 존재하지 않습니다."}
+
+        if (data.user.uid != userID)
+            throw MeasurementDataPermissionException()
+
         repo.deleteById(id)
     }
-
+    @Transactional(readOnly = true)
     override fun findMetricDatasByPeriodAndDate(userId:Long, metric: Metric, period: Period, date: Date): List<MetricData> {
         return repo
             .findMetricDatasByPeriodAndDate(userId,metric,period,date)
